Debugged return of unparsed messages. Fixed the bug that outgoing messages weren't being modified. Added kludge to include MIME headers to subject line of returning message. Got rid of newline at end of tied STDERR. No longer clearing tables by default, added --clear-tables CLI option.
git-svn-id: svn://98.245.80.252/returntome_repository@10 a56a1379-e578-43dd-b219-e15619f59adb

+++ returntome/backup.pl
@@ -47,9 +47,16 @@ $smtp->to($to . "\n");
 $smtp->data();
 $smtp->datasend($str);
 $smtp->dataend();
-$smtp->quit;
 
+my $smtp_response = $smtp->message;
+unless ($smtp_response =~ /2.0.0 OK/) {
+    die "Error: message not sent!";
+} 
+$smtp->quit;
 #clean up:
 unlink 'r2m.tar.gz';
 
-#TODO: check if message was sent sucessfully?
+#now update the subversion repository:
+system "svn commit -F trunk/commit.log";
+system "rm trunk/commit.log";
+system "touch trunk/commit.log";
